requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bootstrap
email-validator
gunicorn
Flask-WTF
python-dotenv
Flask-Migrate
celery
markdown
bleach
wtforms_components
python-slugify
